Refactor large files (>1000 lines) into smaller files under 700 lines
