+++ src/components/TrackSegment/hooks/populateSide.js
@@ -1,0 +1,8 @@
+
+import { populateSidePart1 } from './populateSidePart1';
+import { populateSidePart2 } from './populateSidePart2';
+
+export function populateSide(args) {
+    populateSidePart1(args);
+    populateSidePart2(args);
+}
