+++ src/vehicles/RunnerVehicle.tsx
@@ -1,3 +1,4 @@
+import { useRunnerControls } from './RunnerVehicle/hooks/useRunnerControls';
 import React, { useRef, useEffect, forwardRef, useImperativeHandle } from 'react';
 import { useFrame, useThree } from '@react-three/fiber';
 import { RigidBody, CapsuleCollider, useRapier } from '@react-three/rapier';
@@ -5,188 +6,18 @@ import { usePlayerControls } from '../hooks/usePlayerControls';
 import * as THREE from 'three';
 import { RunnerVehicle as RunnerVehicleClass, SurfaceMaterial, MATERIAL_FROM_BIOME } from '../systems/VehicleSystem';
 import { CollisionParticles } from '../components/CollisionParticles';
-import { getAudioManager, AudioManager } from '../systems/AudioSystem';
 import { WATER_LEVEL, PLAYER_SPAWN, MOVEMENT, PHYSICS } from '../constants/game';
 import { VEHICLE_TUNING } from '../constants/vehicleTuning';
 import { isFloatingPlatform } from '../systems/FloatingObjectRegistry';
 import { useGameStore } from '../systems/GameState';
 
-// Slope detection configuration
-const RAYCAST_ORIGIN_OFFSET = 0.5;
-const RAYCAST_DISTANCE = 5.0;
-const SMOOTHING_FACTOR = 5.0;
-const DEG_TO_RAD = Math.PI / 180;
-
-// Jump state machine configuration
-const JUMP_CONFIG = {
-  FORCE: VEHICLE_TUNING.jumpForce,
-  DOUBLE_JUMP_FORCE: VEHICLE_TUNING.doubleJumpForce,
-  COMMIT_DURATION: 0.1,      // 0.1s strafe disable on jump
-  RECOVERY_DURATION: 0.3,    // 0.3s recovery after landing
-  HIGH_IMPACT_THRESHOLD: 5,  // units/s vertical velocity for camera shake
-  GROUND_CHECK_DIST: 1.5,
-  /** Forward impulse fraction added per unit of sin(slopeAngle) on downhill jumps */
-  SLOPE_FORWARD_BIAS: 0.45,
-  /** Consecutive ungrounded frames required before switching to airborne (smooths bumpy surfaces) */
-  GROUNDED_HYSTERESIS_FRAMES: 3,
-};
-
-// Acceleration multipliers based on slope angle
-const SLOPE_RANGES = {
-  FLAT: { max: 15, multiplier: 1.0 },
-  GENTLE_DOWNSLOPE: { min: 15, max: 45, minMult: 1.0, maxMult: 1.5 },
-  STEEP_DOWNSLOPE: { min: 45, max: 90, multiplier: 1.5 },
-  UPSLOPE: { min: -45, max: 0, minMult: 0.6, maxMult: 0.8 },
-  STEEP_UPSLOPE: { min: -90, max: -45, multiplier: 0.6 },
-};
-
-// Canyon bank riding configuration
-const BANK_CONFIG = {
-  /** Lateral bank angle (°) above which assist forces engage */
-  ASSIST_THRESHOLD: 20,
-  /** Strength of the per-frame "plant" impulse that keeps the player on the bank face */
-  ASSIST_STRENGTH: 22.0,
-  /** Full-assist bank angle (°) – force is linearly interpolated from threshold to here */
-  MAX_BANK_DEG: 70,
-  /** Speed scaling coefficient: plant force grows with velocity (prevents flying off at speed) */
-  SPEED_ASSIST_SCALE: 0.08,
-  /** Anti-roll torque factor applied against X/Z angular velocity on steep banks */
-  ANTI_ROLL_STRENGTH: 6.0,
-  /** Additive speed-multiplier bonus at full bank (gravity-assist on steep canyon walls) */
-  BANK_SPEED_BONUS: 0.3,
-  /** Bank angle (°) at which the jump kick-off XZ bias reaches its maximum (0.5× jump force) */
-  KICKOFF_MAX_BANK_DEG: 60,
-  /** Multiplier applied to SLIDE_FRICTION on steep banks to allow smooth wall-riding */
-  BANK_FRICTION_MULTIPLIER: 2,
-};
-
-const NEAR_MISS_SPEED_THRESHOLD = 12;
-const NEAR_MISS_RAY_LENGTH = 2.6;
-const NEAR_MISS_TOI_MIN = 0.35;
-const NEAR_MISS_TOI_MAX = 1.55;
-
-// Sprint stamina configuration (normalized 0.0–1.0)
-const RUNNER_SPRINT = {
-  /** Stamina drained per second while sprinting on the ground */
-  DRAIN_RATE: 0.25,
-  /** Stamina recovered per second while grounded and NOT sprinting */
-  REGEN_GROUNDED: 0.15,
-  /** Stamina recovered per second while airborne (faster — rewarded air time) */
-  REGEN_AIRBORNE: 0.30,
-  /** Below this value sprint input is rejected (exhaustion lock) */
-  EXHAUSTION_THRESHOLD: 0.0,
-  /** Sprint re-enables once stamina reaches this level (hysteresis) */
-  RECOVERY_THRESHOLD: 0.2,
-  /** Speed multiplier while sprinting */
-  SPEED_MULTIPLIER: 1.5,
-} as const;
-
-// Jump states
-type JumpState = 'grounded' | 'airborne' | 'landing' | 'recovering';
-
-// Dodge states
-type DodgeState = 'ready' | 'dodging' | 'cooldown';
-
-interface DebugImpulse {
-  tag: string;
-  at: number;
-  impulse: { x: number; y: number; z: number };
-}
-
-interface DebugContact {
-  at: number;
-  point: { x: number; y: number; z: number };
-}
-
-interface PhysicsDebugSnapshot {
-  position: { x: number; y: number; z: number };
-  linearVelocity: { x: number; y: number; z: number };
-  angularVelocity: { x: number; y: number; z: number };
-  speed: number;
-  slopeAngle: number;
-  bankAngle: number;
-  isGrounded: boolean;
-  jumpState: JumpState;
-  friction: number;
-  waterfallGravityMultiplier: number;
-  effectiveG: number;
-  extraGravity: number;
-  currentSegmentIndex: number;
-  groundRay: {
-    origin: { x: number; y: number; z: number };
-    hitPoint: { x: number; y: number; z: number } | null;
-    distance: number | null;
-  };
-  groundNormal: { x: number; y: number; z: number };
-  recentImpulses: DebugImpulse[];
-  recentContacts: DebugContact[];
-}
-
-// Audio manager reference
-let audioManager: AudioManager | null = null;
-
-// Initialize audio on first interaction
-const initAudio = () => {
-  if (!audioManager) {
-    // Will be initialized by Experience component
-    audioManager = getAudioManager();
-  }
-};
-
-// Sound helpers
-const playJumpSound = (velocity: number, isDouble: boolean = false) => {
-  initAudio();
-  if (!audioManager) return;
-  
-  // Pitch scales with velocity: 0.8-1.2x
-  const pitch = 0.8 + Math.min(0.4, velocity / 50);
-  const sound = isDouble ? 'double_jump' : 'jump';
-  audioManager.playSound(sound, 1.0, pitch);
-};
-
-const playLandSound = (impactForce: number) => {
-  initAudio();
-  if (!audioManager) return;
-  
-  // Volume scales with force, select sound based on intensity
-  let sound = 'land_soft';
-  let volume = 0.5;
-  
-  if (impactForce > 10) {
-    sound = 'land_impact';
-    volume = Math.min(1.0, impactForce / 15);
-  } else if (impactForce > 5) {
-    sound = 'land_hard';
-    volume = 0.7;
-  }
-  
-  audioManager.playSound(sound, volume, 1.0);
-};
-
-const playFootstep = (material: string, isWet: boolean) => {
-  initAudio();
-  if (!audioManager) return;
-  
-  const sound = AudioManager.getFootstepSound(material, isWet);
-  // Muffled if wet (lower pitch), crisper if dry
-  const pitch = isWet ? 0.85 : 1.0;
-  const volume = isWet ? 0.6 : 0.5;
-  
-  audioManager.playSound(sound, volume, pitch);
-};
-
-const playDodgeSound = () => {
-  initAudio();
-  if (!audioManager) return;
-  audioManager.playSound('dodge', 0.8, 1.0);
-};
-
-// Camera shake effect
-const triggerCameraShake = (intensity: number, duration: number = 0.3) => {
-  window.dispatchEvent(new CustomEvent('camera-shake', {
-    detail: { intensity, duration }
-  }));
-};
+import {
+    RAYCAST_ORIGIN_OFFSET, RAYCAST_DISTANCE, SMOOTHING_FACTOR, DEG_TO_RAD,
+    JUMP_CONFIG, SLOPE_RANGES, BANK_CONFIG, NEAR_MISS_SPEED_THRESHOLD, NEAR_MISS_RAY_LENGTH, NEAR_MISS_TOI_MIN, NEAR_MISS_TOI_MAX, RUNNER_SPRINT, JumpState, DodgeState, DebugImpulse, DebugContact, PhysicsDebugSnapshot
+} from './RunnerVehicle/constants';
+import { initAudio, playJumpSound, playLandSound, playFootstep, playDodgeSound } from './RunnerVehicle/audio';
+import { triggerCameraShake } from './RunnerVehicle/utils';
+import { useRunnerPhysicsState } from './RunnerVehicle/hooks/useRunnerPhysics';
 
 const RunnerVehicle = forwardRef((props, forwardedRef) => {
   const bodyRef = useRef<any>(null);
@@ -194,118 +25,14 @@ const RunnerVehicle = forwardRef((props, forwardedRef) => {
   const { world, rapier } = useRapier();
   const controls = usePlayerControls(camera);
 
-  const vehicle = useRef(new RunnerVehicleClass());
+
   
-  // Slope detection state
-  const slopeState = useRef({
-    currentAngle: 0,
-    targetMultiplier: 1.0,
-    currentMultiplier: 1.0,
-    lastGroundNormal: new THREE.Vector3(0, 1, 0),
-    isGrounded: false,
-    /** Lateral bank angle in degrees; positive = leaning left, negative = leaning right */
-    bankAngle: 0,
-  });
-
-  // Jump state machine
-  const jumpState = useRef({
-    state: 'grounded' as JumpState,
-    hasDoubleJumped: false,
-    airTime: 0,
-    commitTimer: 0,
-    recoveryTimer: 0,
-    lastGroundY: 0,
-    verticalVelocityOnLeave: 0,
-    // Goal 2: Coyote time
-    timeSinceGrounded: 0,
-    // Goal 2: Variable jump height
-    jumpReleased: true,
-  });
-
-  // Grounded hysteresis: counts frames without ground contact before switching to ungrounded
-  const ungroundedFramesRef = useRef(0);
-  
-  // Goal 2: Dodge state machine
-  const dodgeState = useRef({
-    state: 'ready' as DodgeState,
-    timer: 0,
-    direction: new THREE.Vector3(1, 0, 0),
-    nearMissAwarded: false,
-  });
-  
-  // Goal 2: Platform riding state
-  const platformState = useRef({
-    isOnPlatform: false,
-    platformVelocity: new THREE.Vector3(0, 0, 0),
-    platformBody: null as any,
-  });
-
-  // Track last applied gravity multiplier to avoid redundant world.gravity mutations
-  const appliedGravMultRef = useRef(1.0);
-
-  // Sprint stamina hysteresis: when stamina hits 0, sprint is locked out until RECOVERY_THRESHOLD
-  const sprintLockedRef = useRef(false);
-
-  // Reusable Vector3 for camera forward direction (avoids per-frame heap allocations)
-  const jumpForwardDirRef = useRef(new THREE.Vector3());
-
-  // Footstep tracking
-  const footstepState = useRef({
-    distanceTraveled: 0,
-    lastStepDistance: 0,
-    stepInterval: 2.5, // Units between steps
-  });
-
-  // Previous frame state for transition detection
-  const prevFrame = useRef({
-    wasGrounded: true,
-    jumpPressed: false,
-    dodgePressed: false,
-    velocity: new THREE.Vector3(),
-  });
-
-  // Collision & material state
-  const collisionState = useRef({
-    currentBiome: 'summer' as string,
-    activeParticles: [] as Array<{
-      id: number;
-      material: SurfaceMaterial;
-      position: THREE.Vector3;
-      intensity: number;
-    }>,
-  });
-
-  // Goal 2: Collision groups for i-frames
-  const defaultCollisionGroups = useRef(0);
-  const fovRef = useRef(75);
-  const debugState = useRef({
-    recentImpulses: [] as DebugImpulse[],
-    recentContacts: [] as DebugContact[],
-    friction: 0.04,
-  });
-  const debugSnapshotRef = useRef<PhysicsDebugSnapshot>({
-    position: { x: 0, y: 0, z: 0 },
-    linearVelocity: { x: 0, y: 0, z: 0 },
-    angularVelocity: { x: 0, y: 0, z: 0 },
-    speed: 0,
-    slopeAngle: 0,
-    bankAngle: 0,
-    isGrounded: false,
-    jumpState: 'grounded',
-    friction: 0.04,
-    waterfallGravityMultiplier: 1,
-    effectiveG: PHYSICS.GRAVITY,
-    extraGravity: 0,
-    currentSegmentIndex: 0,
-    groundRay: {
-      origin: { x: 0, y: 0, z: 0 },
-      hitPoint: null,
-      distance: null,
-    },
-    groundNormal: { x: 0, y: 1, z: 0 },
-    recentImpulses: [],
-    recentContacts: [],
-  });
+  const state = useRunnerPhysicsState();
+  const {
+    vehicle, slopeState, jumpState, ungroundedFramesRef, dodgeState, platformState,
+    appliedGravMultRef, sprintLockedRef, jumpForwardDirRef, footstepState,
+    prevFrame, collisionState, defaultCollisionGroups, fovRef, debugState, debugSnapshotRef
+  } = state;
 
   useImperativeHandle(forwardedRef, () => bodyRef.current);
 
@@ -475,720 +202,10 @@ const RunnerVehicle = forwardRef((props, forwardedRef) => {
     return 1.0;
   };
 
-  useFrame((state, delta) => {
-    const body = bodyRef.current;
-    if (!body || !world) return;
 
-    const pos = body.translation();
-    const vel = body.linvel();
-    const dt = Math.min(delta, 0.05); // Cap delta for stability
-    const now = performance.now();
-    const frameImpulses = new Map<string, { x: number; y: number; z: number }>();
-    const applyImpulseWithDebugTracking = (tag: string, impulse: { x: number; y: number; z: number }) => {
-      body.applyImpulse(impulse, true);
-      const prev = frameImpulses.get(tag);
-      if (prev) {
-        prev.x += impulse.x;
-        prev.y += impulse.y;
-        prev.z += impulse.z;
-      } else {
-        frameImpulses.set(tag, { x: impulse.x, y: impulse.y, z: impulse.z });
-      }
-    };
-
-    // === SLOPE DETECTION ===
-    const slopeAngle = calculateSlopeAngle();
-    
-    const groundRay = new rapier.Ray(
-      { x: pos.x, y: pos.y + RAYCAST_ORIGIN_OFFSET, z: pos.z },
-      { x: 0, y: -1, z: 0 }
-    );
-    const groundHit = world.castRay(groundRay, RAYCAST_DISTANCE, true);
-    const groundRayDistance = groundHit
-      ? (typeof (groundHit as any).timeOfImpact === 'function'
-          ? (groundHit as any).timeOfImpact()
-          : groundHit.timeOfImpact)
-      : null;
-    const groundRayHitPoint = groundRayDistance !== null
-      ? {
-          x: pos.x,
-          y: pos.y + RAYCAST_ORIGIN_OFFSET - groundRayDistance,
-          z: pos.z,
-        }
-      : null;
-    if (typeof (groundRay as any).free === 'function') (groundRay as any).free();
-
-    // Hysteresis: require GROUNDED_HYSTERESIS_FRAMES consecutive misses before going airborne.
-    // This prevents spurious state changes when the player skims over small rocks/lips.
-    const rawGrounded = !!groundHit;
-    if (rawGrounded) {
-      ungroundedFramesRef.current = 0;
-    } else {
-      // Cap at threshold so the counter doesn't grow unbounded while airborne
-      ungroundedFramesRef.current = Math.min(
-        ungroundedFramesRef.current + 1,
-        JUMP_CONFIG.GROUNDED_HYSTERESIS_FRAMES
-      );
-    }
-    const isGrounded = rawGrounded || ungroundedFramesRef.current < JUMP_CONFIG.GROUNDED_HYSTERESIS_FRAMES;
-    slopeState.current.isGrounded = isGrounded;
-
-    // Goal 2: Platform detection via raycast handle registry
-    platformState.current.isOnPlatform = false;
-    platformState.current.platformBody = null;
-    if (groundHit) {
-      try {
-        const collider = groundHit.collider;
-        const parent = collider?.parent?.();
-        if (parent) {
-          if (isFloatingPlatform(parent.handle)) {
-            platformState.current.isOnPlatform = true;
-            platformState.current.platformBody = true; // used only as boolean at line 709
-            const pVel = parent.linvel();
-            platformState.current.platformVelocity.set(pVel.x, pVel.y, pVel.z);
-          }
-          if (typeof (parent as any).free === 'function') (parent as any).free();
-        }
-      } catch {
-        // Ignore colliders without parent
-      }
-      if (typeof (groundHit as any).free === 'function') (groundHit as any).free();
-    }
-    
-    if (isGrounded) {
-      slopeState.current.currentAngle = slopeAngle;
-      // Base pitch multiplier
-      let baseMult = calculateSlopeMultiplier(slopeAngle);
-      // Bank bonus: steep canyon walls contribute an additional flow/speed boost
-      const absBankDeg = Math.abs(slopeState.current.bankAngle);
-      if (absBankDeg > BANK_CONFIG.ASSIST_THRESHOLD) {
-        const bankT = Math.min(1.0, (absBankDeg - BANK_CONFIG.ASSIST_THRESHOLD) /
-                                     (BANK_CONFIG.MAX_BANK_DEG - BANK_CONFIG.ASSIST_THRESHOLD));
-        baseMult = Math.min(1.5, baseMult + bankT * BANK_CONFIG.BANK_SPEED_BONUS);
-      }
-      slopeState.current.targetMultiplier = baseMult;
-    } else {
-      slopeState.current.targetMultiplier = THREE.MathUtils.lerp(
-        slopeState.current.targetMultiplier,
-        1.0,
-        dt * 2.0
-      );
-    }
-
-    slopeState.current.currentMultiplier = THREE.MathUtils.lerp(
-      slopeState.current.currentMultiplier,
-      slopeState.current.targetMultiplier,
-      Math.min(1.0, dt * SMOOTHING_FACTOR)
-    );
-
-    // === BANK ASSIST (canyon U-wall riding) ===
-    // When the player is on a steep lateral bank, apply two forces:
-    //  1. A "plant" impulse directed into the bank face (opposite horizontal normal component)
-    //     that acts like centripetal force and grows with speed so the player stays planted.
-    //  2. Anti-roll torque impulses that damp capsule tumbling around the X and Z world axes.
-    const bankAngle = slopeState.current.bankAngle;
-    const absBankDeg = Math.abs(bankAngle);
-    if (isGrounded && absBankDeg > BANK_CONFIG.ASSIST_THRESHOLD) {
-      const bankT = Math.min(1.0, (absBankDeg - BANK_CONFIG.ASSIST_THRESHOLD) /
-                                   (BANK_CONFIG.MAX_BANK_DEG - BANK_CONFIG.ASSIST_THRESHOLD));
-      const gn = slopeState.current.lastGroundNormal;
-      const speed2D = Math.sqrt(vel.x * vel.x + vel.z * vel.z);
-      const speedFactor = 1.0 + speed2D * BANK_CONFIG.SPEED_ASSIST_SCALE;
-      const plantMag = BANK_CONFIG.ASSIST_STRENGTH * bankT * speedFactor * dt;
-      // Push player into bank face (negate horizontal normal → toward wall surface)
-      applyImpulseWithDebugTracking('bankAssist', { x: -gn.x * plantMag, y: 0, z: -gn.z * plantMag });
-      // Damp capsule angular velocity on X and Z to prevent tumbling on steep walls
-      const angVel = body.angvel();
-      const antiRoll = BANK_CONFIG.ANTI_ROLL_STRENGTH * bankT * dt;
-      body.applyTorqueImpulse({
-        x: -angVel.x * antiRoll,
-        y: 0,
-        z: -angVel.z * antiRoll,
-      }, true);
-    }
-
-    // === GRAVITY MULTIPLIER (waterfall / steep rapids) ===
-    // Apply waterfallGravityMultiplier from Zustand to the Rapier world without
-    // subscribing to the store (avoiding a React re-render every frame).
-    const gravMultCurrent = useGameStore.getState().waterfallGravityMultiplier;
-    if (gravMultCurrent !== appliedGravMultRef.current) {
-      appliedGravMultRef.current = gravMultCurrent;
-      // Mutate in-place to avoid an unnecessary object allocation
-      world.gravity.x = 0;
-      world.gravity.y = PHYSICS.GRAVITY * gravMultCurrent;
-      world.gravity.z = 0;
-    }
-
-    // === CAMERA FORWARD DIRECTION (used for slope-biased jump and dodge) ===
-    const jumpForwardDir = jumpForwardDirRef.current;
-    camera.getWorldDirection(jumpForwardDir);
-    jumpForwardDir.y = 0;
-    if (jumpForwardDir.lengthSq() > 0.001) jumpForwardDir.normalize();
-
-    // Pre-compute slope trig used by both grounded and coyote-time jump paths
-    const slopeRad = Math.abs(slopeState.current.currentAngle) * DEG_TO_RAD;
-    const sinSlope = Math.sin(slopeRad);
-
-    // === JUMP STATE MACHINE ===
-    const js = jumpState.current;
-    const { jump, leftward, rightward, dodge, sprint } = controls.getControls();
-    const jumpJustPressed = jump && !prevFrame.current.jumpPressed;
-    const dodgeJustPressed = dodge && !prevFrame.current.dodgePressed;
-
-    // === SPRINT STAMINA (Zustand-backed, no stale closures) ===
-    // Read via getState() to avoid stale closure; never use the hook form in useFrame.
-    const storeState = useGameStore.getState();
-    let stamina = storeState.sprintStamina;
-
-    // Hysteresis: once exhausted, lock sprint until RECOVERY_THRESHOLD is reached
-    if (stamina <= RUNNER_SPRINT.EXHAUSTION_THRESHOLD) {
-      sprintLockedRef.current = true;
-    } else if (sprintLockedRef.current && stamina >= RUNNER_SPRINT.RECOVERY_THRESHOLD) {
-      sprintLockedRef.current = false;
-    }
-
-    // Airborne sprint intent is ignored — no drain, faster recovery instead
-    const isAirborne = js.state === 'airborne';
-    const isSprinting = sprint && !sprintLockedRef.current && !isAirborne;
-
-    if (isSprinting) {
-      // Drain while grounded and sprinting
-      stamina = Math.max(0, stamina - RUNNER_SPRINT.DRAIN_RATE * dt);
-    } else if (isAirborne) {
-      // Faster recovery while airborne
-      stamina = Math.min(1, stamina + RUNNER_SPRINT.REGEN_AIRBORNE * dt);
-    } else {
-      // Normal grounded recovery when not sprinting
-      stamina = Math.min(1, stamina + RUNNER_SPRINT.REGEN_GROUNDED * dt);
-    }
-    storeState.setSprintStamina(stamina);
-
-    // Goal 2: Update coyote time
-    if (isGrounded) {
-      js.timeSinceGrounded = 0;
-    } else {
-      js.timeSinceGrounded += dt;
-    }
-
-    // State transitions
-    switch (js.state) {
-      case 'grounded': {
-        if (!isGrounded) {
-          // Left ground - became airborne
-          js.state = 'airborne';
-          js.hasDoubleJumped = false;
-          js.airTime = 0;
-          js.verticalVelocityOnLeave = vel.y;
-          js.lastGroundY = pos.y;
-        } else if (jumpJustPressed) {
-          // Initiate jump with forward slope bias so downhill jumps carry momentum
-          const jumpForce = JUMP_CONFIG.FORCE * Math.max(0.8, slopeState.current.currentMultiplier);
-          // Add a forward component proportional to slope steepness
-          const forwardBias = jumpForce * sinSlope * JUMP_CONFIG.SLOPE_FORWARD_BIAS;
-          // On steep banks, blend a kick-off component along the surface normal's XZ plane
-          // so the player bounces away from the wall rather than straight up.
-          const gn = slopeState.current.lastGroundNormal;
-          const bankBlend = Math.min(0.5, Math.abs(slopeState.current.bankAngle) / BANK_CONFIG.KICKOFF_MAX_BANK_DEG);
-          applyImpulseWithDebugTracking('jump', {
-            x: jumpForwardDir.x * forwardBias + gn.x * jumpForce * bankBlend,
-            y: jumpForce,
-            z: jumpForwardDir.z * forwardBias + gn.z * jumpForce * bankBlend,
-          });
-          
-          js.state = 'airborne';
-          js.hasDoubleJumped = false;
-          js.airTime = 0;
-          js.commitTimer = JUMP_CONFIG.COMMIT_DURATION;
-          js.verticalVelocityOnLeave = jumpForce;
-          js.jumpReleased = false;
-          
-          // F1: Play jump sound with velocity-scaled pitch
-          playJumpSound(Math.sqrt(vel.x * vel.x + vel.z * vel.z));
-        }
-        break;
-      }
-
-      case 'airborne': {
-        js.airTime += dt;
-        
-        // Commit phase - disable strafe
-        if (js.commitTimer > 0) {
-          js.commitTimer -= dt;
-        }
-
-        // Goal 2: Variable jump height - early release cuts upward velocity
-        if (!jump && !js.jumpReleased && vel.y > 0) {
-          js.jumpReleased = true;
-          body.setLinvel({
-            x: vel.x,
-            y: vel.y * MOVEMENT.JUMP_CUT_MULTIPLIER,
-            z: vel.z
-          }, true);
-        }
-
-        // Goal 2: Coyote time jump - if we just left ground, still allow jump
-        if (jumpJustPressed && js.timeSinceGrounded <= MOVEMENT.COYOTE_TIME && js.airTime < 0.15 && !js.hasDoubleJumped) {
-          const jumpForce = JUMP_CONFIG.FORCE * 0.9;
-          const forwardBias = jumpForce * sinSlope * JUMP_CONFIG.SLOPE_FORWARD_BIAS;
-          // Apply same bank kick-off bias as regular jump
-          const gn = slopeState.current.lastGroundNormal;
-          const bankBlend = Math.min(0.5, Math.abs(slopeState.current.bankAngle) / BANK_CONFIG.KICKOFF_MAX_BANK_DEG);
-          applyImpulseWithDebugTracking('coyoteJump', {
-            x: jumpForwardDir.x * forwardBias + gn.x * jumpForce * bankBlend,
-            y: jumpForce,
-            z: jumpForwardDir.z * forwardBias + gn.z * jumpForce * bankBlend,
-          });
-          js.hasDoubleJumped = false;
-          js.commitTimer = JUMP_CONFIG.COMMIT_DURATION;
-          js.jumpReleased = false;
-          playJumpSound(Math.sqrt(vel.x * vel.x + vel.z * vel.z));
-          // Remain airborne but reset timers
-          js.airTime = 0;
-          js.timeSinceGrounded = 999;
-        }
-
-        // Double jump
-        if (jumpJustPressed && !js.hasDoubleJumped && js.airTime > 0.05) {
-          const doubleJumpForce = JUMP_CONFIG.DOUBLE_JUMP_FORCE;
-          // Cancel some downward velocity for responsive feel
-          const newVelY = Math.max(vel.y * 0.3, 0) + doubleJumpForce;
-          body.setLinvel({ x: vel.x, y: newVelY, z: vel.z }, true);
-          
-          js.hasDoubleJumped = true;
-          js.jumpReleased = false;
-          // F1: Play double jump sound
-          playJumpSound(Math.sqrt(vel.x * vel.x + vel.z * vel.z), true);
-        }
-
-        // Landing detection
-        if (isGrounded && js.airTime > 0.05) {
-          js.state = 'landing';
-          
-          // Calculate impact velocity
-          const impactVel = Math.abs(vel.y);
-          
-          // F1: Play landing sound scaled by impact force
-          playLandSound(impactVel);
-          
-          // High impact landing check
-          if (impactVel > JUMP_CONFIG.HIGH_IMPACT_THRESHOLD) {
-            const shakeIntensity = Math.min(1.0, (impactVel - JUMP_CONFIG.HIGH_IMPACT_THRESHOLD) / 10);
-            triggerCameraShake(shakeIntensity);
-          }
-        }
-        break;
-      }
-
-      case 'landing': {
-        // Immediate transition to recovery
-        js.state = 'recovering';
-        js.recoveryTimer = JUMP_CONFIG.RECOVERY_DURATION;
-        break;
-      }
-
-      case 'recovering': {
-        js.recoveryTimer -= dt;
-        
-        if (js.recoveryTimer <= 0) {
-          if (isGrounded) {
-            js.state = 'grounded';
-          } else {
-            js.state = 'airborne';
-          }
-        }
-        break;
-      }
-    }
-
-    // === DODGE STATE MACHINE (Goal 2) ===
-    const ds = dodgeState.current;
-    switch (ds.state) {
-      case 'ready': {
-        if (dodgeJustPressed && isGrounded) {
-          ds.state = 'dodging';
-          ds.timer = MOVEMENT.DODGE_DURATION;
-          ds.nearMissAwarded = false;
-          
-          // Determine dodge direction from input
-          const forwardDir = new THREE.Vector3();
-          camera.getWorldDirection(forwardDir);
-          forwardDir.y = 0;
-          forwardDir.normalize();
-          const rightDir = new THREE.Vector3();
-          rightDir.crossVectors(forwardDir, camera.up).normalize();
-          
-          ds.direction.set(0, 0, 0);
-          if (leftward) {
-            ds.direction.sub(rightDir);
-          } else if (rightward) {
-            ds.direction.add(rightDir);
-          } else if (controls.getControls().backward) {
-            ds.direction.sub(forwardDir);
-          } else {
-            ds.direction.add(forwardDir);
-          }
-          if (ds.direction.lengthSq() < 0.001) {
-            ds.direction.copy(rightDir); // Default dodge right
-          }
-          ds.direction.normalize();
-          
-          // Apply dodge impulse
-          applyImpulseWithDebugTracking('dodge', {
-            x: ds.direction.x * MOVEMENT.DODGE_FORCE,
-            y: 2.0, // Slight upward lift
-            z: ds.direction.z * MOVEMENT.DODGE_FORCE
-          });
-          
-          // I-frames: temporarily disable collision with obstacles (not ground)
-          // We achieve this by applying a collision group mask if supported
-          try {
-            // Rapier doesn't have simple collision group toggling on RigidBody ref,
-            // so we use a flag that other systems can check
-            body.userData = { ...body.userData, isDodging: true };
-          } catch (e) { /* ignore */ }
-          
-          playDodgeSound();
-          triggerCameraShake(0.3, 0.15);
-          
-          // Broadcast dodge state for HUD
-          window.dispatchEvent(new CustomEvent('player-dodge', { detail: { state: 'start' } }));
-          useGameStore.getState().setIsDodging(true);
-        }
-        break;
-      }
-      
-      case 'dodging': {
-        const dodgeSpeed = Math.sqrt(vel.x * vel.x + vel.z * vel.z);
-        if (!ds.nearMissAwarded && dodgeSpeed >= NEAR_MISS_SPEED_THRESHOLD) {
-          const sideDir = new THREE.Vector3(-ds.direction.z, 0, ds.direction.x);
-          if (sideDir.lengthSq() > 0.001) {
-            sideDir.normalize();
-            const rayOrigin = { x: pos.x, y: pos.y + 0.6, z: pos.z };
-            const checkSide = (sign: number) => {
-              const dir = sideDir.clone().multiplyScalar(sign);
-              const ray = new rapier.Ray(rayOrigin, { x: dir.x, y: 0, z: dir.z });
-              const hit = world.castRay(ray, NEAR_MISS_RAY_LENGTH, true);
-              const toi = hit
-                ? (typeof (hit as any).timeOfImpact === 'function'
-                  ? (hit as any).timeOfImpact()
-                  : (hit as any).timeOfImpact)
-                : null;
-              if (hit && typeof (hit as any).free === 'function') (hit as any).free();
-              return toi !== null && toi >= NEAR_MISS_TOI_MIN && toi <= NEAR_MISS_TOI_MAX;
-            };
-
-            if (checkSide(1) || checkSide(-1)) {
-              ds.nearMissAwarded = true;
-              window.dispatchEvent(new CustomEvent('player-near-miss', { detail: { speed: dodgeSpeed } }));
-            }
-          }
-        }
-
-        ds.timer -= dt;
-        if (ds.timer <= 0) {
-          ds.state = 'cooldown';
-          ds.timer = MOVEMENT.DODGE_COOLDOWN;
-          try {
-            body.userData = { ...body.userData, isDodging: false };
-          } catch (e) { /* ignore */ }
-          window.dispatchEvent(new CustomEvent('player-dodge', { detail: { state: 'end' } }));
-          useGameStore.getState().setIsDodging(false);
-        }
-        break;
-      }
-      
-      case 'cooldown': {
-        ds.timer -= dt;
-        if (ds.timer <= 0) {
-          ds.state = 'ready';
-        }
-        break;
-      }
-    }
-
-    // === COLLISION DETECTION ===
-    // Check for sudden velocity changes indicating collision
-    const prevVelocity = prevFrame.current.velocity;
-    const velocityDelta = new THREE.Vector3(vel.x - prevVelocity.x, vel.y - prevVelocity.y, vel.z - prevVelocity.z);
-    const impactForce = velocityDelta.length() / dt;
-    
-    // Detect high-impact collision (skip if dodging with i-frames)
-    const isDodging = ds.state === 'dodging' && ds.timer > (MOVEMENT.DODGE_DURATION - MOVEMENT.DODGE_I_FRAMES);
-    if (impactForce > 8 && !isGrounded && !isDodging) {
-      const material = MATERIAL_FROM_BIOME[collisionState.current.currentBiome] || SurfaceMaterial.ROCK;
-      const contactPoint = new THREE.Vector3(pos.x, pos.y - 0.5, pos.z);
-      
-      vehicle.current.processCollision(
-        material,
-        impactForce,
-        contactPoint,
-        new THREE.Vector3(vel.x, vel.y, vel.z)
-      );
-      
-      // Add visual particles for high impact
-      if (impactForce > vehicle.current['highImpactThreshold']) {
-        const newParticle = {
-          id: Date.now(),
-          material,
-          position: contactPoint.clone(),
-          intensity: Math.min(1, impactForce / 20),
-        };
-        collisionState.current.activeParticles.push(newParticle);
-        debugState.current.recentContacts.push({
-          at: now,
-          point: { x: contactPoint.x, y: contactPoint.y, z: contactPoint.z },
-        });
-      }
-    }
-
-    // === FOOTSTEP AUDIO (F1) ===
-    if (isGrounded && js.state === 'grounded') {
-      const speed = Math.sqrt(vel.x * vel.x + vel.z * vel.z);
-      if (speed > 0.5) {
-        footstepState.current.distanceTraveled += speed * dt;
-        
-        if (footstepState.current.distanceTraveled - footstepState.current.lastStepDistance >= footstepState.current.stepInterval) {
-          footstepState.current.lastStepDistance = footstepState.current.distanceTraveled;
-          
-          // Determine material and wetness
-          const material = collisionState.current.currentBiome.includes('autumn') ? 'moss' : 'rock';
-          const isWet = pos.y < WATER_LEVEL + 1.0;
-          
-          playFootstep(material, isWet);
-        }
-      }
-    }
-
-    // Store previous frame state
-    prevFrame.current.wasGrounded = isGrounded;
-    prevFrame.current.jumpPressed = jump;
-    prevFrame.current.dodgePressed = dodge;
-    prevFrame.current.velocity.set(vel.x, vel.y, vel.z);
-
-    // === MOVEMENT INPUT ===
-    const { forward, backward } = controls.getControls();
-
-    // Camera-relative direction
-    const forwardDir = new THREE.Vector3();
-    camera.getWorldDirection(forwardDir);
-    forwardDir.y = 0;
-    forwardDir.normalize();
-
-    const rightDir = new THREE.Vector3();
-    rightDir.crossVectors(forwardDir, camera.up).normalize();
-
-    // === APPLY FORCES ===
-    const flowResponsiveness = VEHICLE_TUNING.flowResponsiveness;
-    const flowMultiplier = slopeState.current.currentMultiplier;
-    
-    // Recovery state: reduced turn speed
-    const recoveryFactor = js.state === 'recovering' 
-      ? 0.5  // 50% speed during recovery
-      : 1.0;
-
-    applyImpulseWithDebugTracking('flow', { x: 0, y: 0, z: -flowResponsiveness * flowMultiplier * dt });
-
-    const baseSpeed = VEHICLE_TUNING.baseSpeed;
-    const sprintMultiplier = isSprinting ? RUNNER_SPRINT.SPEED_MULTIPLIER : 1.0;
-    const speed = baseSpeed * flowMultiplier * recoveryFactor * sprintMultiplier;
-
-    if (forward) {
-      applyImpulseWithDebugTracking('forwardInput', {
-        x: forwardDir.x * speed * dt,
-        y: 0,
-        z: forwardDir.z * speed * dt
-      });
-    }
-    if (backward) {
-      applyImpulseWithDebugTracking('backwardInput', {
-        x: forwardDir.x * -speed * 0.6 * dt,
-        y: 0,
-        z: forwardDir.z * -speed * 0.6 * dt
-      });
-    }
-
-    // Strafe with commit/recovery restrictions
-    const inCommitPhase = js.state === 'airborne' && js.commitTimer > 0;
-    const canStrafe = !inCommitPhase && js.state !== 'recovering' && ds.state !== 'dodging';
-
-    if (canStrafe) {
-      if (leftward) {
-        applyImpulseWithDebugTracking('leftwardInput', {
-          x: rightDir.x * -speed * 0.8 * dt,
-          y: 0,
-          z: rightDir.z * -speed * 0.8 * dt
-        });
-      }
-      if (rightward) {
-        applyImpulseWithDebugTracking('rightwardInput', {
-          x: rightDir.x * speed * 0.8 * dt,
-          y: 0,
-          z: rightDir.z * speed * 0.8 * dt
-        });
-      }
-    }
-
-    // Goal 2: Helper to set friction on all colliders
-    const setBodyFriction = (frictionValue: number) => {
-      debugState.current.friction = frictionValue;
-      try {
-        const n = body.numColliders();
-        for (let i = 0; i < n; i++) {
-          body.collider(i).setFriction(frictionValue);
-        }
-      } catch (e) {
-        // Fallback: friction not adjustable on this body
-      }
-    };
-
-    // === SLOPE-AWARE HORIZONTAL SPEED CAP ===
-    // Reward momentum on steep downhill terrain while preventing physics explosions.
-    // Effective cap = maxHorizontalSpeed * (1 + slopeBonusScale * (flowMultiplier - 1))
-    // so steep sections scale the limit proportionally to slopeBonusScale.
-    const effectiveMaxH = VEHICLE_TUNING.maxHorizontalSpeed *
-      (1 + VEHICLE_TUNING.slopeBonusScale * (flowMultiplier - 1));
-    const horizSpeed = Math.sqrt(vel.x * vel.x + vel.z * vel.z);
-    if (horizSpeed > effectiveMaxH) {
-      const scale = effectiveMaxH / horizSpeed;
-      body.setLinvel({ x: vel.x * scale, y: vel.y, z: vel.z * scale }, true);
-    }
-
-    // Goal 2: Slide mechanic — activate on steep forward pitch OR steep lateral bank
-    const isSlideInput = controls.getControls().brake &&
-      (slopeAngle > MOVEMENT.SLIDE_MIN_SLOPE ||
-       Math.abs(slopeState.current.bankAngle) > MOVEMENT.SLIDE_MIN_SLOPE);
-    if (isSlideInput && isGrounded) {
-      // Reduce friction during slide for speed boost; on a bank the slide runs along the wall
-      setBodyFriction(MOVEMENT.SLIDE_FRICTION);
-      const slideBoost = MOVEMENT.SLIDE_SPEED_BOOST;
-      applyImpulseWithDebugTracking('slideBoost', {
-        x: forwardDir.x * speed * (slideBoost - 1.0) * dt,
-        y: 0,
-        z: forwardDir.z * speed * (slideBoost - 1.0) * dt
-      });
-    } else {
-      // Restore friction based on current material
-      // On steep banks, use lower friction to allow smooth wall-riding
-      const baseFriction = vehicle.current.getConfig().friction;
-      const absBankForFriction = Math.abs(slopeState.current.bankAngle);
-      const friction = absBankForFriction > BANK_CONFIG.ASSIST_THRESHOLD
-        ? THREE.MathUtils.lerp(
-            baseFriction,
-            MOVEMENT.SLIDE_FRICTION * BANK_CONFIG.BANK_FRICTION_MULTIPLIER,
-            Math.min(1.0, (absBankForFriction - BANK_CONFIG.ASSIST_THRESHOLD) /
-                          (BANK_CONFIG.MAX_BANK_DEG - BANK_CONFIG.ASSIST_THRESHOLD))
-          )
-        : baseFriction;
-      setBodyFriction(friction);
-    }
-
-    // Goal 2: Platform riding - apply platform velocity to player
-    if (platformState.current.isOnPlatform && platformState.current.platformBody) {
-      const pVel = platformState.current.platformVelocity;
-      const transfer = MOVEMENT.PLAYER_MOMENTUM_TRANSFER ?? 0.3;
-      applyImpulseWithDebugTracking('platformTransfer', {
-        x: (pVel.x - vel.x) * transfer,
-        y: (pVel.y - vel.y) * transfer * 0.5, // Less vertical transfer
-        z: (pVel.z - vel.z) * transfer
-      });
-    }
-
-    // === IN-WATER PHYSICS ===
-    // Shallow-water running: lower damping, lateral current push.
-    const inShallowWater = pos.y < WATER_LEVEL + 1.0;
-    if (inShallowWater) {
-      try {
-        body.setLinearDamping(VEHICLE_TUNING.linearDampingShallowWater);
-      } catch (_e) { /* damping setter unavailable */ }
-      const w = window as unknown as { __watershedFlowSpeed?: number };
-      const segFlow = (typeof window !== 'undefined' && Number.isFinite(w.__watershedFlowSpeed))
-        ? (w.__watershedFlowSpeed as number)
-        : 1.0;
-      const currentPush = 0.3 * segFlow * dt;
-      applyImpulseWithDebugTracking('waterCurrent', {
-        x: forwardDir.x * currentPush,
-        y: 0,
-        z: forwardDir.z * currentPush,
-      });
-    } else {
-      try {
-        body.setLinearDamping(VEHICLE_TUNING.linearDampingAir);
-      } catch (_e) { /* damping setter unavailable */ }
-    }
-
-    // Camera follow (first-person, smooth)
-    // Guard against NaN from Rapier during physics init — lerping NaN permanently corrupts camera matrix
-    if (isFinite(pos.x) && isFinite(pos.y) && isFinite(pos.z)) {
-      const targetPos = new THREE.Vector3(pos.x, pos.y + 1.65, pos.z);
-      camera.position.lerp(targetPos, 0.12);
-    }
-
-    // FOV kick — speed-based expansion + waterfall punch
-    {
-      const segIdx = useGameStore.getState().currentSegmentIndex;
-      const isWaterfallSeg = segIdx === 14 || segIdx === 29;
-      // Horizontal speed only — vertical drops shouldn't affect FOV alone
-      const hSpeed = Math.sqrt(vel.x * vel.x + vel.z * vel.z);
-      const targetFov = 75 + Math.min(12, hSpeed * 0.45) + (isWaterfallSeg ? 10 : 0);
-      fovRef.current += (targetFov - fovRef.current) * (1 - Math.exp(-delta * 5));
-      if (Math.abs(camera.fov - fovRef.current) > 0.05) {
-        camera.fov = fovRef.current;
-        camera.updateProjectionMatrix();
-      }
-    }
-
-    frameImpulses.forEach((impulse, tag) => {
-      const mag = Math.sqrt(impulse.x * impulse.x + impulse.y * impulse.y + impulse.z * impulse.z);
-      if (mag < 0.001) return;
-      debugState.current.recentImpulses.push({ tag, at: now, impulse });
-    });
-    debugState.current.recentImpulses = debugState.current.recentImpulses
-      .filter((entry) => now - entry.at <= 2000)
-      .slice(-16);
-    debugState.current.recentContacts = debugState.current.recentContacts
-      .filter((entry) => now - entry.at <= 2500)
-      .slice(-8);
-
-    const gravMult = useGameStore.getState().waterfallGravityMultiplier;
-    const snapshot = debugSnapshotRef.current;
-    snapshot.position = { x: pos.x, y: pos.y, z: pos.z };
-    snapshot.linearVelocity = { x: vel.x, y: vel.y, z: vel.z };
-    const angVel = body.angvel();
-    snapshot.angularVelocity = { x: angVel.x, y: angVel.y, z: angVel.z };
-    snapshot.speed = Math.sqrt(vel.x * vel.x + vel.y * vel.y + vel.z * vel.z);
-    snapshot.slopeAngle = slopeState.current.currentAngle;
-    snapshot.bankAngle = slopeState.current.bankAngle;
-    snapshot.isGrounded = isGrounded;
-    snapshot.jumpState = js.state;
-    snapshot.friction = debugState.current.friction;
-    snapshot.waterfallGravityMultiplier = gravMultCurrent;
-    snapshot.effectiveG = PHYSICS.GRAVITY * gravMultCurrent;
-    snapshot.extraGravity = PHYSICS.GRAVITY * (gravMultCurrent - 1);
-    snapshot.currentSegmentIndex = useGameStore.getState().currentSegmentIndex;
-    snapshot.groundRay = {
-      origin: { x: pos.x, y: pos.y + RAYCAST_ORIGIN_OFFSET, z: pos.z },
-      hitPoint: groundRayHitPoint,
-      distance: groundRayDistance,
-    };
-    snapshot.groundNormal = {
-      x: slopeState.current.lastGroundNormal.x,
-      y: slopeState.current.lastGroundNormal.y,
-      z: slopeState.current.lastGroundNormal.z,
-    };
-    snapshot.recentImpulses = debugState.current.recentImpulses;
-    snapshot.recentContacts = debugState.current.recentContacts;
-
-    const bodyUserData = (body.userData ??= {});
-    bodyUserData.physicsDebug = snapshot;
-    if (typeof window !== 'undefined') {
-      (window as any).__watershedPhysicsDebug = snapshot;
-    }
+  useRunnerControls({
+    bodyRef, camera, controls, vehicleState: state
   });
-
-  // Remove completed particles
-  const removeParticle = (id: number) => {
-    collisionState.current.activeParticles = collisionState.current.activeParticles.filter(p => p.id !== id);
-  };
 
   return (
     <>
