+++ src/vehicles/RunnerVehicle/hooks/useRunnerControls.ts
@@ -1,0 +1,14 @@
+import { useFrame, useThree } from '@react-three/fiber';
+import { useRapier } from '@react-three/rapier';
+import { updateRunnerPhysics } from './RunnerPhysicsStep';
+
+export function useRunnerControls({ bodyRef, camera, controls, vehicleState }) {
+  const { world, rapier } = useRapier();
+
+  useFrame((state, delta) => {
+    if (!bodyRef.current) return;
+    updateRunnerPhysics({
+        state, delta, world, body: bodyRef.current, rapier, camera, controls, vehicleState
+    });
+  });
+}
