+++ src/vehicles/RunnerVehicle/utils.ts
@@ -1,0 +1,6 @@
+// Camera shake effect
+export const triggerCameraShake = (intensity: number, duration: number = 0.3) => {
+  window.dispatchEvent(new CustomEvent('camera-shake', {
+    detail: { intensity, duration }
+  }));
+};
