--- webpack.config.js
@@ -1,23 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  entry: './src/main.ts',
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        use: 'ts-loader',
-        exclude: /node_modules/,
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.ts', '.js'],
-  },
-  output: {
-    filename: 'main.bundle.js',
-    path: path.resolve(__dirname, 'dist'),
-  },
-  mode: 'development',
-  watch: true
-};
